+++ src/main/java/kitchenpos/application/TableService.java
@@ -3,7 +3,6 @@ package kitchenpos.application;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.dao.OrderRepository;
 import kitchenpos.dao.OrderTableRepository;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
@@ -11,6 +10,7 @@ import kitchenpos.dto.request.TableCreateRequest;
 import kitchenpos.dto.request.TableUpdateEmptyRequest;
 import kitchenpos.dto.request.TableUpdateGuestRequest;
 import kitchenpos.dto.response.TableResponse;
+import kitchenpos.order.repository.OrderRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
