+++ src/main/java/kitchenpos/order/dto/response/OrderResponse.java
@@ -1,8 +1,8 @@
-package kitchenpos.dto.response;
+package kitchenpos.order.dto.response;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Order;
+import kitchenpos.order.domain.Order;
 
 public class OrderResponse {
